requirements.txt
streamlit>=1.30.0
fitparse>=1.2.0
pandas>=2.1.0
plotly>=6.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
fpdf2>=2.7.0
fpdf2
matplotlib